[Bronze V] Title: A+B - 9, Time: 0 ms, Memory: 1112 KB, Score: 75 point -BaekjoonHub

+++ 백준/Bronze/15740. A＋B － 9/A＋B － 9.c
@@ -2,7 +2,7 @@
 
 int main()
 {
-    int a, b;
-    scanf("%d %d", &a, &b);
-    printf("%d", a + b);
+    long long int a, b;
+    scanf("%lld %lld", &a, &b);
+    printf("%lld", a + b);
 }
